requirements.txt
cycler==0.11.0
fonttools==4.43.0
kiwisolver==1.4.5
numpy==1.26.1
opencv-python==4.8.1.78
packaging==21.3
Pillow==10.1.0
pyautogui==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.8
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pywin32==306
pywinauto==0.6.8
six==1.16.0
